Split csm_camera into csm_io + csm_analysis and collapse cam1/cam2 plotting
Breaks the 1541-LOC function-based csm_camera.py into three layers:

- csm_io.py: the ASP-mirrored camera-model readers (toCsmPixel, read_csm_cam,
  read_tsai_cam, read_angles, ...), kept function-based with a provenance
  header noting they are synced from upstream orbit_plot.py.
- csm_analysis.py: the asp_plot-specific analysis (get_orbit_plot_gdf,
  reproject_ecef, poly_fit), built on csm_io.
- csm_camera.py: the plotting layer. The near-verbatim cam1/cam2 halves of
  csm_camera_summary_plot (~210 lines duplicated by index) collapse into a
  single _plot_camera() called once per camera, and the duplicated linescan
  x-axis tick logic into _apply_frame_xaxis().

csm_camera.py re-exports the moved symbols for backward compatibility with
notebooks and downstream imports; stereo_geometry.py now imports
estim_satellite_orientation from csm_io directly. Figure output is unchanged,
verified by the golden line-digest characterization tests added previously.

Closes #131.
